Unify trim tests so that they're together
