+++ rust/kcl-lib/src/frontend/trim.rs
@@ -4358,766 +4358,3 @@ pub(crate) async fn execute_trim_operations_simple(
     scene_graph_delta.invalidates_ids = invalidates_ids;
     Ok((source_delta, scene_graph_delta))
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_is_point_on_line_segment_exactly_on_segment() {
-        let point = Coords2d { x: 5.0, y: 5.0 };
-        let segment_start = Coords2d { x: 0.0, y: 0.0 };
-        let segment_end = Coords2d { x: 10.0, y: 10.0 };
-        let result = is_point_on_line_segment(point, segment_start, segment_end, EPSILON_POINT_ON_SEGMENT);
-        assert!(result.is_some());
-        if let Some(r) = result {
-            assert!((r.x - 5.0).abs() < 1e-5);
-            assert!((r.y - 5.0).abs() < 1e-5);
-        }
-    }
-
-    #[test]
-    fn test_is_point_on_line_segment_at_start() {
-        let point = Coords2d { x: 0.0, y: 0.0 };
-        let segment_start = Coords2d { x: 0.0, y: 0.0 };
-        let segment_end = Coords2d { x: 10.0, y: 10.0 };
-        let result = is_point_on_line_segment(point, segment_start, segment_end, EPSILON_POINT_ON_SEGMENT);
-        assert!(result.is_some());
-        if let Some(r) = result {
-            assert_eq!(r.x, 0.0);
-            assert_eq!(r.y, 0.0);
-        }
-    }
-
-    #[test]
-    fn test_is_point_on_line_segment_at_end() {
-        let point = Coords2d { x: 10.0, y: 10.0 };
-        let segment_start = Coords2d { x: 0.0, y: 0.0 };
-        let segment_end = Coords2d { x: 10.0, y: 10.0 };
-        let result = is_point_on_line_segment(point, segment_start, segment_end, EPSILON_POINT_ON_SEGMENT);
-        assert!(result.is_some());
-        if let Some(r) = result {
-            assert_eq!(r.x, 10.0);
-            assert_eq!(r.y, 10.0);
-        }
-    }
-
-    #[test]
-    fn test_is_point_on_line_segment_before_start() {
-        let point = Coords2d { x: -1.0, y: -1.0 };
-        let segment_start = Coords2d { x: 0.0, y: 0.0 };
-        let segment_end = Coords2d { x: 10.0, y: 10.0 };
-        let result = is_point_on_line_segment(point, segment_start, segment_end, EPSILON_POINT_ON_SEGMENT);
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_is_point_on_line_segment_after_end() {
-        let point = Coords2d { x: 11.0, y: 11.0 };
-        let segment_start = Coords2d { x: 0.0, y: 0.0 };
-        let segment_end = Coords2d { x: 10.0, y: 10.0 };
-        let result = is_point_on_line_segment(point, segment_start, segment_end, EPSILON_POINT_ON_SEGMENT);
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_is_point_on_line_segment_off_to_side() {
-        let point = Coords2d { x: 5.0, y: 6.0 };
-        let segment_start = Coords2d { x: 0.0, y: 0.0 };
-        let segment_end = Coords2d { x: 10.0, y: 10.0 };
-        let result = is_point_on_line_segment(point, segment_start, segment_end, EPSILON_POINT_ON_SEGMENT);
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_is_point_on_line_segment_horizontal() {
-        let point = Coords2d { x: 5.0, y: 0.0 };
-        let segment_start = Coords2d { x: 0.0, y: 0.0 };
-        let segment_end = Coords2d { x: 10.0, y: 0.0 };
-        let result = is_point_on_line_segment(point, segment_start, segment_end, EPSILON_POINT_ON_SEGMENT);
-        assert!(result.is_some());
-        if let Some(r) = result {
-            assert_eq!(r.x, 5.0);
-            assert_eq!(r.y, 0.0);
-        }
-    }
-
-    #[test]
-    fn test_is_point_on_line_segment_vertical() {
-        let point = Coords2d { x: 0.0, y: 5.0 };
-        let segment_start = Coords2d { x: 0.0, y: 0.0 };
-        let segment_end = Coords2d { x: 0.0, y: 10.0 };
-        let result = is_point_on_line_segment(point, segment_start, segment_end, EPSILON_POINT_ON_SEGMENT);
-        assert!(result.is_some());
-        if let Some(r) = result {
-            assert_eq!(r.x, 0.0);
-            assert_eq!(r.y, 5.0);
-        }
-    }
-
-    #[test]
-    fn test_is_point_on_line_segment_degenerate() {
-        let point = Coords2d { x: 0.0, y: 0.0 };
-        let segment_start = Coords2d { x: 0.0, y: 0.0 };
-        let segment_end = Coords2d { x: 0.0, y: 0.0 };
-        let result = is_point_on_line_segment(point, segment_start, segment_end, EPSILON_POINT_ON_SEGMENT);
-        assert!(result.is_some());
-        if let Some(r) = result {
-            assert_eq!(r.x, 0.0);
-            assert_eq!(r.y, 0.0);
-        }
-    }
-
-    #[test]
-    fn test_is_point_on_line_segment_far_from_degenerate() {
-        let point = Coords2d { x: 1.0, y: 1.0 };
-        let segment_start = Coords2d { x: 0.0, y: 0.0 };
-        let segment_end = Coords2d { x: 0.0, y: 0.0 };
-        let result = is_point_on_line_segment(point, segment_start, segment_end, EPSILON_POINT_ON_SEGMENT);
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_line_segment_intersection_crossing_segments() {
-        let result = line_segment_intersection(
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 10.0, y: 10.0 },
-            Coords2d { x: 0.0, y: 10.0 },
-            Coords2d { x: 10.0, y: 0.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result.is_some());
-        if let Some(r) = result {
-            assert!((r.x - 5.0).abs() < 1e-5);
-            assert!((r.y - 5.0).abs() < 1e-5);
-        }
-    }
-
-    #[test]
-    fn test_line_segment_intersection_shared_endpoint() {
-        let result = line_segment_intersection(
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 5.0, y: 5.0 },
-            Coords2d { x: 5.0, y: 5.0 },
-            Coords2d { x: 10.0, y: 10.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result.is_some());
-        if let Some(r) = result {
-            assert!((r.x - 5.0).abs() < 1e-5);
-            assert!((r.y - 5.0).abs() < 1e-5);
-        }
-    }
-
-    #[test]
-    fn test_line_segment_intersection_parallel() {
-        let result = line_segment_intersection(
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 10.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 5.0 },
-            Coords2d { x: 10.0, y: 5.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_line_segment_intersection_non_intersecting() {
-        let result = line_segment_intersection(
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 5.0, y: 5.0 },
-            Coords2d { x: 10.0, y: 10.0 },
-            Coords2d { x: 15.0, y: 15.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_line_segment_intersection_outside_segments() {
-        let result = line_segment_intersection(
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 5.0, y: 5.0 },
-            Coords2d { x: 10.0, y: 10.0 },
-            Coords2d { x: 20.0, y: 20.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_line_segment_intersection_t_intersection() {
-        let result = line_segment_intersection(
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 10.0, y: 0.0 },
-            Coords2d { x: 5.0, y: -5.0 },
-            Coords2d { x: 5.0, y: 5.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result.is_some());
-        if let Some(r) = result {
-            assert!((r.x - 5.0).abs() < 1e-5);
-            assert!((r.y - 0.0).abs() < 1e-5);
-        }
-    }
-
-    #[test]
-    fn test_project_point_onto_segment_at_start() {
-        let result = project_point_onto_segment(
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 10.0, y: 10.0 },
-        );
-        assert!((result - 0.0).abs() < 1e-5);
-    }
-
-    #[test]
-    fn test_project_point_onto_segment_at_end() {
-        let result = project_point_onto_segment(
-            Coords2d { x: 10.0, y: 10.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 10.0, y: 10.0 },
-        );
-        assert!((result - 1.0).abs() < 1e-5);
-    }
-
-    #[test]
-    fn test_project_point_onto_segment_at_midpoint() {
-        let result = project_point_onto_segment(
-            Coords2d { x: 5.0, y: 5.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 10.0, y: 10.0 },
-        );
-        assert!((result - 0.5).abs() < 1e-5);
-    }
-
-    #[test]
-    fn test_project_point_onto_segment_before_start() {
-        let result = project_point_onto_segment(
-            Coords2d { x: -1.0, y: -1.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 10.0, y: 10.0 },
-        );
-        assert!(result < 0.0);
-    }
-
-    #[test]
-    fn test_project_point_onto_segment_after_end() {
-        let result = project_point_onto_segment(
-            Coords2d { x: 11.0, y: 11.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 10.0, y: 10.0 },
-        );
-        assert!(result > 1.0);
-    }
-
-    #[test]
-    fn test_project_point_onto_segment_horizontal() {
-        let result = project_point_onto_segment(
-            Coords2d { x: 5.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 10.0, y: 0.0 },
-        );
-        assert!((result - 0.5).abs() < 1e-5);
-    }
-
-    #[test]
-    fn test_project_point_onto_segment_vertical() {
-        let result = project_point_onto_segment(
-            Coords2d { x: 0.0, y: 5.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 10.0 },
-        );
-        assert!((result - 0.5).abs() < 1e-5);
-    }
-
-    #[test]
-    fn test_perpendicular_distance_to_segment_on_segment() {
-        let result = perpendicular_distance_to_segment(
-            Coords2d { x: 5.0, y: 5.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 10.0, y: 10.0 },
-        );
-        assert!((result - 0.0).abs() < 1e-5);
-    }
-
-    #[test]
-    fn test_perpendicular_distance_to_segment_perpendicular() {
-        // Point at [5, 5] with segment from [0, 0] to [10, 0]
-        // Perpendicular distance should be 5
-        let result = perpendicular_distance_to_segment(
-            Coords2d { x: 5.0, y: 5.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 10.0, y: 0.0 },
-        );
-        assert!((result - 5.0).abs() < 1e-5);
-    }
-
-    #[test]
-    fn test_perpendicular_distance_to_segment_outside_segment() {
-        // Point at [-1, 0] with segment from [0, 0] to [10, 0]
-        // Should return distance to start point (1)
-        let result = perpendicular_distance_to_segment(
-            Coords2d { x: -1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 10.0, y: 0.0 },
-        );
-        assert!((result - 1.0).abs() < 1e-5);
-    }
-
-    #[test]
-    fn test_perpendicular_distance_to_segment_degenerate() {
-        // Degenerate segment (start == end)
-        let result = perpendicular_distance_to_segment(
-            Coords2d { x: 1.0, y: 1.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-        );
-        // Distance should be sqrt(2)
-        assert!((result - 2.0_f64.sqrt()).abs() < 1e-5);
-    }
-
-    #[test]
-    fn test_is_point_on_arc_on_arc() {
-        // Arc from [1, 0] to [0, 1] with center at [0, 0] (quarter circle)
-        // Use a point that's exactly on the unit circle
-        let angle = std::f64::consts::PI / 4.0; // 45 degrees
-        let point = Coords2d {
-            x: libm::cos(angle),
-            y: libm::sin(angle),
-        };
-        let result = is_point_on_arc(
-            point,
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 1.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result);
-    }
-
-    #[test]
-    fn test_is_point_on_arc_at_start() {
-        let result = is_point_on_arc(
-            Coords2d { x: 1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 1.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result);
-    }
-
-    #[test]
-    fn test_is_point_on_arc_at_end() {
-        let result = is_point_on_arc(
-            Coords2d { x: 0.0, y: 1.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 1.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result);
-    }
-
-    #[test]
-    fn test_is_point_on_arc_off_arc() {
-        // Point on circle but not on the arc (opposite side)
-        let result = is_point_on_arc(
-            Coords2d { x: -1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 1.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(!result);
-    }
-
-    #[test]
-    fn test_is_point_on_arc_not_on_circle() {
-        let result = is_point_on_arc(
-            Coords2d { x: 2.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 1.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(!result);
-    }
-
-    #[test]
-    fn test_is_point_on_arc_wraps_around() {
-        // Arc from 350° to 10° (wraps around)
-        let center = Coords2d { x: 0.0, y: 0.0 };
-        let start = Coords2d {
-            x: libm::cos(350.0 * std::f64::consts::PI / 180.0),
-            y: libm::sin(350.0 * std::f64::consts::PI / 180.0),
-        };
-        let end = Coords2d {
-            x: libm::cos(10.0 * std::f64::consts::PI / 180.0),
-            y: libm::sin(10.0 * std::f64::consts::PI / 180.0),
-        };
-        let point = Coords2d {
-            x: libm::cos(5.0 * std::f64::consts::PI / 180.0),
-            y: libm::sin(5.0 * std::f64::consts::PI / 180.0),
-        };
-        let result = is_point_on_arc(point, center, start, end, EPSILON_POINT_ON_SEGMENT);
-        assert!(result);
-    }
-
-    #[test]
-    fn test_line_arc_intersection_intersects() {
-        // Line from [-2, 0] to [2, 0], arc from [1, 0] to [0, 1] with center at [0, 0]
-        // Should intersect at [1, 0] (arc start)
-        let result = line_arc_intersection(
-            Coords2d { x: -2.0, y: 0.0 },
-            Coords2d { x: 2.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 1.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result.is_some());
-        if let Some(r) = result {
-            assert!((r.x - 1.0).abs() < 1e-5);
-            assert!((r.y - 0.0).abs() < 1e-5);
-        }
-    }
-
-    #[test]
-    fn test_line_arc_intersection_no_intersection() {
-        // Line that doesn't intersect the arc
-        let result = line_arc_intersection(
-            Coords2d { x: 0.0, y: 2.0 },
-            Coords2d { x: 0.0, y: 3.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 1.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result.is_none());
-    }
-
-    #[test]
-    fn test_line_arc_intersection_tangent() {
-        // Line tangent to arc (touches at one point)
-        // Line from [0, 1] to [0, 2], arc from [1, 0] to [0, 1] with center at [0, 0]
-        // Should intersect at [0, 1] (arc end)
-        let result = line_arc_intersection(
-            Coords2d { x: 0.0, y: 1.0 },
-            Coords2d { x: 0.0, y: 2.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 1.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result.is_some());
-        if let Some(r) = result {
-            assert!((r.x - 0.0).abs() < 1e-5);
-            assert!((r.y - 1.0).abs() < 1e-5);
-        }
-    }
-
-    #[test]
-    fn test_line_arc_intersection_horizontal_line() {
-        // Horizontal line at y=0.5 intersecting arc from [1,0] to [0,1] centered at [0,0]
-        let result = line_arc_intersection(
-            Coords2d { x: -1.0, y: 0.5 },
-            Coords2d { x: 1.0, y: 0.5 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 1.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result.is_some());
-        if let Some(r) = result {
-            // Should intersect at approximately [0.866, 0.5] or [-0.866, 0.5]
-            // But only [0.866, 0.5] is on the arc from [1,0] to [0,1]
-            assert!(r.x.abs() > 0.8);
-            assert!((r.y - 0.5).abs() < 0.1);
-        }
-    }
-
-    #[test]
-    fn test_line_arc_intersection_endpoint() {
-        // Line starts at arc endpoint
-        let result = line_arc_intersection(
-            Coords2d { x: 1.0, y: 0.0 },
-            Coords2d { x: 2.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 0.0 },
-            Coords2d { x: 1.0, y: 0.0 },
-            Coords2d { x: 0.0, y: 1.0 },
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        assert!(result.is_some());
-        if let Some(r) = result {
-            assert!((r.x - 1.0).abs() < 1e-5);
-            assert!((r.y - 0.0).abs() < 1e-5);
-        }
-    }
-
-    #[test]
-    fn test_arc_arc_intersection() {
-        // Test case matching TypeScript test: two arcs that may or may not intersect
-        // arc1: center [0, 0], start [1, 0], end [0, 1] (quarter circle from 0° to 90°)
-        // arc2: center [1, 0], start [2, 0], end [1, 1] (quarter circle from 0° to 90°)
-        let result = arc_arc_intersection(
-            Coords2d { x: 0.0, y: 0.0 }, // arc1 center
-            Coords2d { x: 1.0, y: 0.0 }, // arc1 start
-            Coords2d { x: 0.0, y: 1.0 }, // arc1 end
-            Coords2d { x: 1.0, y: 0.0 }, // arc2 center
-            Coords2d { x: 2.0, y: 0.0 }, // arc2 start
-            Coords2d { x: 1.0, y: 1.0 }, // arc2 end
-            EPSILON_POINT_ON_SEGMENT,
-        );
-        // arc_arc_intersection may return None if no intersection, or Some(point)
-        // The test just verifies the function works without panicking
-        // In this case, the arcs may or may not intersect depending on geometry
-        // The important thing is that the function returns Option<Coords2d>
-        assert!(result.is_none() || result.is_some());
-    }
-
-    #[test]
-    fn test_get_next_trim_spawn_line_intersection() {
-        use serde_json::json;
-
-        // Create a simple line segment object
-        let line_obj_json = json!({
-            "id": 0,
-            "kind": {
-                "type": "Segment",
-                "segment": {
-                    "type": "Line",
-                    "start": 1,
-                    "end": 2,
-                    "ctor": {
-                        "type": "Line",
-                        "start": { "x": { "type": "Number", "value": 0.0, "units": "None" }, "y": { "type": "Number", "value": 0.0, "units": "None" } },
-                        "end": { "x": { "type": "Number", "value": 10.0, "units": "None" }, "y": { "type": "Number", "value": 10.0, "units": "None" } }
-                    },
-                    "ctor_applicable": false,
-                    "construction": false
-                }
-            },
-            "label": "",
-            "comments": "",
-            "artifact_id": "00000000-0000-0000-0000-000000000000",
-            "source": { "type": "Simple", "range": [0, 0, 0] }
-        });
-
-        // Create point objects for start and end
-        let start_point_json = json!({
-            "id": 1,
-            "kind": {
-                "type": "Segment",
-                "segment": {
-                    "type": "Point",
-                    "position": {
-                        "x": { "type": "Number", "value": 0.0, "units": "None" },
-                        "y": { "type": "Number", "value": 0.0, "units": "None" }
-                    },
-                    "ctor": null,
-                    "owner": null,
-                    "freedom": "Free",
-                    "constraints": []
-                }
-            },
-            "label": "",
-            "comments": "",
-            "artifact_id": "00000000-0000-0000-0000-000000000000",
-            "source": { "type": "Simple", "range": [0, 0, 0] }
-        });
-
-        let end_point_json = json!({
-            "id": 2,
-            "kind": {
-                "type": "Segment",
-                "segment": {
-                    "type": "Point",
-                    "position": {
-                        "x": { "type": "Number", "value": 10.0, "units": "None" },
-                        "y": { "type": "Number", "value": 10.0, "units": "None" }
-                    },
-                    "ctor": null,
-                    "owner": null,
-                    "freedom": "Free",
-                    "constraints": []
-                }
-            },
-            "label": "",
-            "comments": "",
-            "artifact_id": "00000000-0000-0000-0000-000000000000",
-            "source": { "type": "Simple", "range": [0, 0, 0] }
-        });
-
-        // Deserialize JSON into Object types
-        let line_obj: Object = serde_json::from_value(line_obj_json).unwrap();
-        let start_point: Object = serde_json::from_value(start_point_json).unwrap();
-        let end_point: Object = serde_json::from_value(end_point_json).unwrap();
-
-        let objects = vec![line_obj, start_point, end_point];
-
-        // Trim line that intersects: from [0, 10] to [10, 0]
-        let points = vec![Coords2d { x: 0.0, y: 10.0 }, Coords2d { x: 10.0, y: 0.0 }];
-
-        let result = get_next_trim_spawn(&points, 0, &objects, UnitLength::Millimeters);
-
-        match result {
-            TrimItem::Spawn { trim_spawn_coords, .. } => {
-                // Should intersect at [5, 5]
-                assert!((trim_spawn_coords.x - 5.0).abs() < 1e-5);
-                assert!((trim_spawn_coords.y - 5.0).abs() < 1e-5);
-            }
-            TrimItem::None { .. } => {
-                panic!("Expected intersection but got None");
-            }
-        }
-    }
-
-    #[test]
-    fn test_get_next_trim_spawn_no_intersection() {
-        use serde_json::json;
-
-        // Create a line segment that won't intersect
-        let line_obj_json = json!({
-            "id": 0,
-            "kind": {
-                "type": "Segment",
-                "segment": {
-                    "type": "Line",
-                    "start": 1,
-                    "end": 2,
-                    "ctor": {
-                        "type": "Line",
-                        "start": { "x": { "type": "Number", "value": 0.0, "units": "None" }, "y": { "type": "Number", "value": 0.0, "units": "None" } },
-                        "end": { "x": { "type": "Number", "value": 10.0, "units": "None" }, "y": { "type": "Number", "value": 0.0, "units": "None" } }
-                    },
-                    "ctor_applicable": false,
-                    "construction": false
-                }
-            },
-            "label": "",
-            "comments": "",
-            "artifact_id": "00000000-0000-0000-0000-000000000000",
-            "source": { "type": "Simple", "range": [0, 0, 0] }
-        });
-
-        let start_point_json = json!({
-            "id": 1,
-            "kind": {
-                "type": "Segment",
-                "segment": {
-                    "type": "Point",
-                    "position": {
-                        "x": { "type": "Number", "value": 0.0, "units": "None" },
-                        "y": { "type": "Number", "value": 0.0, "units": "None" }
-                    },
-                    "ctor": null,
-                    "owner": null,
-                    "freedom": "Free",
-                    "constraints": []
-                }
-            },
-            "label": "",
-            "comments": "",
-            "artifact_id": "00000000-0000-0000-0000-000000000000",
-            "source": { "type": "Simple", "range": [0, 0, 0] }
-        });
-
-        let end_point_json = json!({
-            "id": 2,
-            "kind": {
-                "type": "Segment",
-                "segment": {
-                    "type": "Point",
-                    "position": {
-                        "x": { "type": "Number", "value": 10.0, "units": "None" },
-                        "y": { "type": "Number", "value": 0.0, "units": "None" }
-                    },
-                    "ctor": null,
-                    "owner": null,
-                    "freedom": "Free",
-                    "constraints": []
-                }
-            },
-            "label": "",
-            "comments": "",
-            "artifact_id": "00000000-0000-0000-0000-000000000000",
-            "source": { "type": "Simple", "range": [0, 0, 0] }
-        });
-
-        // Deserialize JSON into Object types
-        let line_obj: Object = serde_json::from_value(line_obj_json).unwrap();
-        let start_point: Object = serde_json::from_value(start_point_json).unwrap();
-        let end_point: Object = serde_json::from_value(end_point_json).unwrap();
-
-        let objects = vec![line_obj, start_point, end_point];
-
-        // Trim line that doesn't intersect: from [0, 10] to [10, 10]
-        let points = vec![Coords2d { x: 0.0, y: 10.0 }, Coords2d { x: 10.0, y: 10.0 }];
-
-        let result = get_next_trim_spawn(&points, 0, &objects, UnitLength::Millimeters);
-
-        match result {
-            TrimItem::None { .. } => {
-                // Expected
-            }
-            TrimItem::Spawn { .. } => {
-                panic!("Expected no intersection but got Spawn");
-            }
-        }
-    }
-
-    /// Helper function to execute the full trim flow and return the resulting KCL code
-    /// This uses the public `execute_trim_flow` function from the module
-    async fn execute_trim_flow(
-        kcl_code: &str,
-        trim_points: &[Coords2d],
-        sketch_id: ObjectId,
-    ) -> Result<super::TrimFlowResult, String> {
-        // Use the public execute_trim_flow function
-        super::execute_trim_flow(kcl_code, trim_points, sketch_id).await
-    }
-
-    #[tokio::test]
-    async fn test_execute_trim_flow_infrastructure() {
-        // Simple test to verify the infrastructure works
-        // This is a minimal test that just verifies we can parse KCL and set up the frontend
-        let kcl_code = r#"
-@settings(experimentalFeatures = allow)
-
-sketch(on = YZ) {
-  line1 = sketch2::line(start = [var 0mm, var 5mm], end = [var 0mm, var -5mm])
-  line2 = sketch2::line(start = [var 5mm, var 0mm], end = [var -5mm, var 0mm])
-}
-"#;
-
-        let trim_points = vec![Coords2d { x: -2.0, y: -2.0 }, Coords2d { x: -2.0, y: 2.0 }];
-
-        // This should at least parse and set up the frontend without errors
-        // The actual trim might not work yet if operations aren't fully implemented
-        let result = execute_trim_flow(kcl_code, &trim_points, ObjectId(0)).await;
-
-        // For now, just verify it doesn't panic
-        // Once operations are fully implemented, we can add assertions
-        match result {
-            Ok(_) => {
-                // Success - infrastructure is working
-            }
-            Err(e) => {
-                // This is expected for now since some operations aren't implemented
-                // Just verify we got a reasonable error message
-                // The error could be various things like parsing errors, execution errors, or empty results
-                assert!(
-                    e.contains("Failed")
-                        || e.contains("not yet implemented")
-                        || e.contains("Reached max iterations")
-                        || e.contains("No trim operations")
-                        || e.contains("No operations were executed")
-                        || e.contains("parse")
-                        || e.contains("execute")
-                );
-            }
-        }
-    }
-}
